+++ src/constants/settings.ts
@@ -1,20 +1,22 @@
 export type SettingsStorage = {
-  version: number
+  version: 1.1
   lists: {
     common: boolean
     fakeNews: boolean
     gambling: boolean
     social: boolean
+    ipGrabbers: boolean
   }
 }
 
 export const DEFAULT_SETTINGS: SettingsStorage = {
-  version: 1,
+  version: 1.1,
   lists: {
     common: true,
     fakeNews: false,
     gambling: false,
     social: false,
+    ipGrabbers: false,
   },
 }