+++ CHANGELOG.md
@@ -10,11 +10,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - 💬 List for ip grabbers
+- 💄 Minor UI changes to improve the feel of the extension
 
 ### Changed
 
 - ⚡️ Move cliqz engine loading to a webworker to improve responsiveness
+- ⚡️ Cosmetics load at document start to improve responsiveness
 - 🔧 Additional lists (social, gambling and fakenews) are disabled by default
+- 💄 Shield icon now fits in with browser theme
 
 ### Fixed